backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
praw==7.7.1
motor==3.3.2
pymongo==4.6.0
openai>=1.6.1
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
textblob>=0.17.0
celery>=5.3.0
redis>=5.0.0
websockets>=12.0
jsonschema>=4.20.0
loguru>=0.7.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
aiofiles>=23.2.0
httpx>=0.25.0
beautifulsoup4>=4.12.0